% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{assess_stability}
\alias{assess_stability}
\title{Function to assess the stability of a given model with respect to the ensemble matrix}
\usage{
assess_stability(SEEK.out, vocabulary, K, n.top.words = 10, font.size = 8)
}
\arguments{
\item{SEEK.out}{Output of run_SEEK.}

\item{vocabulary}{Vector of vocabulary in the order corresponding to the input corpus matrix.}

\item{K}{K of interest.}

\item{n.top.words}{Number of top words per topic to analyze (default is 10).}
}
\value{
A pheatmap object and the stability measures of the block of top words from each topic.
}
\description{
Function to assess the stability of a given model with respect to the ensemble matrix
}
\examples{
assess_stability(SEEK.out, vocabulary, 6, n.top.words = 10)
}
