% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{plot_results}
\alias{plot_results}
\title{Function to plot heatmap of results}
\usage{
plot_results(SEEK.out, vocabulary, filter.value = 0)
}
\arguments{
\item{SEEK.out}{Output of run_SEEK.}

\item{vocabulary}{Vector of vocabulary in the order corresponding to the input corpus matrix.}

\item{filter.value}{Vocabulary words with no off-diagonal value greater than or equal to filter.value will be omitted (default is 0)}
}
\value{
A pheatmap object.
}
\description{
Function to plot heatmap of results
}
\examples{
plot_results(SEEK.out, vocabulary, filter.value = 0.5)
}
