% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_SEEK}
\alias{run_SEEK}
\title{Main function to run SEEK-VEC}
\usage{
run_SEEK(D, Ks, method = "ts", threshold = 0.1)
}
\arguments{
\item{D}{A vocabulary x documents count matrix.}

\item{Ks}{List of candidate values of K.}

\item{method}{Topic modeling method to use - one of c("ts" (default),"nmf","ctm","lda").}

\item{threshold}{A proportion between 0 and 1 or a number of top words per topic (default is 0.1).}
}
\value{
A SEEK.out object containing the input Ks; the topic, loadings, hallmark, and joint hallmark projection matrices for each value of K; the normalized eigenvector corresponding to the consensus matrix, and the ensembled joint hallmark projection matrix O.
}
\description{
Main function to run SEEK-VEC
}
\examples{
run_SEEK(D, 3:12, method = "ts", threshold = 0.2)
run_SEEK(D, 3:12, method = "ts", threshold = 10)
}
