% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_seekvfi}
\alias{run_seekvfi}
\title{Main function to run SEEK-VFI}
\usage{
run_seekvfi(counts, Ks, ns = 1, im = 100, seed = 1)
}
\arguments{
\item{Ks}{List of candidate values of K.}

\item{ns}{start parameter for kmeans (default 1)}

\item{im}{iter.max parameter for kmeans (default 100)}

\item{seed}{Random seed (default 1)}

\item{D}{A genes x cells expression count matrix with gene names as the rownames.}
}
\value{
A data frame containing the SEEK-VFI scores of each gene
}
\description{
Main function to run SEEK-VFI
}
\examples{
run_seekvfi(D, 3:12)
}
