% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_seekvfi}
\alias{run_seekvfi}
\title{Main function to run SEEK-VFI}
\usage{
run_seekvfi(D, Ks)
}
\arguments{
\item{D}{A genes x cells expression count matrix with gene names as the rownames.}

\item{Ks}{List of candidate values of K.}
}
\value{
A data frame containing the SEEK-VFI scores of each gene
}
\description{
Main function to run SEEK-VFI
}
\examples{
seekvfi(D, 3:12)
}
